refactor: Move blueprints and registry to service
