+++ actor/inventory/inventory.go
@@ -9,8 +9,8 @@ import (
 
 	"github.com/0xa1-red/empires-of-avalon/instrumentation/traces"
 	"github.com/0xa1-red/empires-of-avalon/persistence"
-	"github.com/0xa1-red/empires-of-avalon/pkg/blueprints"
-	"github.com/0xa1-red/empires-of-avalon/pkg/blueprints/registry"
+	"github.com/0xa1-red/empires-of-avalon/pkg/service/blueprints"
+	"github.com/0xa1-red/empires-of-avalon/pkg/service/registry"
 	"github.com/0xa1-red/empires-of-avalon/protobuf"
 	intnats "github.com/0xa1-red/empires-of-avalon/transport/nats"
 	"github.com/asynkron/protoactor-go/cluster"
