+++ pkg/service/blueprints/resource.go
