+++ pkg/service/game/inventory.go
@@ -3,8 +3,8 @@ package game
 import (
 	"context"
 
-	"github.com/0xa1-red/empires-of-avalon/common"
 	"github.com/0xa1-red/empires-of-avalon/gamecluster"
+	"github.com/0xa1-red/empires-of-avalon/pkg/service/blueprints"
 	"github.com/0xa1-red/empires-of-avalon/protobuf"
 	"github.com/google/uuid"
 	"go.opentelemetry.io/otel"
@@ -14,7 +14,7 @@ import (
 )
 
 func Describe(ctx context.Context, userID uuid.UUID) (*protobuf.DescribeInventoryResponse, error) {
-	inventoryID := common.GetInventoryID(userID)
+	inventoryID := blueprints.GetInventoryID(userID)
 
 	slog.Info("getting inventory grain client", "id", inventoryID.String())
 	inventory := protobuf.GetInventoryGrainClient(gamecluster.GetC(), inventoryID.String())
