+++ pkg/service/registry/errors.go
