+++ pkg/service/registry/registry.go
@@ -5,7 +5,7 @@ import (
 	"io"
 	"os"
 
-	"github.com/0xa1-red/empires-of-avalon/pkg/blueprints"
+	"github.com/0xa1-red/empires-of-avalon/pkg/service/blueprints"
 	"github.com/google/uuid"
 	"golang.org/x/exp/slog"
 	"gopkg.in/yaml.v3"
